Sweet Alert

+++ programming-hero/module-56/client/src/components/Coffee.jsx
@@ -1,19 +1,38 @@
 /* eslint-disable react/prop-types */
 import { FaEye, FaPen, FaTrash } from "react-icons/fa";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const Coffee = ({ coffee, coffees, setCoffees }) => {
-
   // handleDelete
   const handleDelete = (id) => {
-    fetch(`http://localhost:5000/coffees/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        setCoffees(coffees.filter((coffee) => coffee._id !== id));
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/coffees/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+            }
+            setCoffees(coffees.filter((coffee) => coffee._id !== id));
+          });
+      }
+    });
   };
 
   const { _id: id, name, chef, price, photo } = coffee;
